XEP-0156: Accept explicitly-allowed Origin for CORS
The Access-Control-Allow-Origin header can contain "*" or alternatively an
Origin value that should be allowed.

There are limitations with "*". For example, when sending "credentials" (which
includes cookies, basic auth, or TLS client certificates), a wildcard is not
permitted.

See https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS/Errors/CORSNotSupportingCredentials

Since Prosody 0.12, Prosody will echo the origin provided by the client (if any),
and the compliance tester would fail because it only looked for "*".

This commit fixes it to accept either "*" or the origin we inserted in the request.